1.0.9

+++ Ais.IO.Command/hash_execute.cpp
@@ -103,7 +103,7 @@ void hash_execute::HashStart(Hashes& hash) {
 	std::vector<unsigned char> output;
 	cryptography_libary::ValueEncode(hash.input_option, hash.Input, input);
 	cryptography_libary::ValueEncode(hash.salt_option, hash.Salt, salt);
-	int length = (hash.Length != 0 && hash.Mode == HASH_TYPE::HASH_SHA3_KE_128) || (hash.Length != 0 && hash.Mode == HASH_TYPE::HASH_SHA3_KE_256)
+	size_t length = (hash.Length != 0 && hash.Mode == HASH_TYPE::HASH_SHA3_KE_128) || (hash.Length != 0 && hash.Mode == HASH_TYPE::HASH_SHA3_KE_256)
 		? hash.Length 
 		: ((GetHashLength)HashFunctions.at("-hash-length"))(hash.Mode);
 	output.resize(length);
